pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.40.1"
fireworks-ai = "^0.15.8"
python-dotenv = "^1.0.1"
openai = "^1.54.5"
pydantic = "^2.9.2"
pytesseract = "^0.3.13"
opencv-python-headless = "^4.10.0.84"
watchdog = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

